workflow: docker-image: Specify platform for debian-arm64
Specify --platform=linux/arm64 for building debian-arm64. Without this,
it is unable to pull from the debian:trixie-slim image, since we are
building for arm64 image for amd64 qemu system.

Since we have only 2 Dockerfiles for now and the foreseeable future,
replace the loop with 2 individual commands to accomplish this.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>